final nits
