+++ src/deepsparse/server/__init__.py
@@ -19,4 +19,4 @@ DeepSparse server implementation to serve ONNX files and pipelines through
 the DeepSparse Engine.
 """
 
-from .cli import main
+# from .cli import main
