+++ src/deepsparse/server/config.py
@@ -15,10 +15,9 @@
 
 from typing import Any, Dict, List, Optional, Tuple, Union
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, validator
 
 from deepsparse import DEEPSPARSE_ENGINE, PipelineConfig
-from deepsparse.loggers.config import MetricFunctionConfig
 from deepsparse.tasks import SupportedTasks
 
 
@@ -47,6 +46,45 @@ class ImageSizesConfig(BaseModel):
     image_sizes: List[Tuple[int, int]] = Field(
         description="The list of image sizes the model should accept"
     )
+
+
+class MetricFunctionConfig(BaseModel):
+    """
+    Holds logging configuration for a metric function
+    """
+
+    func: str = Field(
+        description="The name that specifies the metric function to be applied. "
+        "It can be: "
+        "1) a built-in function name "
+        "2) a dynamic import function of the form "
+        "'<path_to_the_python_script>:<function_name>' "
+        "3) a framework function (e.g. np.mean or torch.mean)"
+    )
+
+    frequency: int = Field(
+        description="Specifies how often the function should be applied"
+        "(measured in numbers of inference calls).",
+        default=1,
+    )
+
+    target_loggers: Optional[List[str]] = Field(
+        default=None,
+        description="Overrides the global logger configuration in "
+        "the context of the DeepSparse server. "
+        "If not None, this configuration stops logging data "
+        "to globally specified loggers, and will only use "
+        "the subset of loggers (specified here by a list of their names).",
+    )
+
+    @validator("frequency")
+    def non_zero_frequency(cls, frequency: int) -> int:
+        if frequency <= 0:
+            raise ValueError(
+                f"Passed frequency: {frequency}, but "
+                "frequency must be a positive integer greater equal 1"
+            )
+        return frequency
 
 
 class EndpointConfig(BaseModel):
